Krissi Miller
Remove a portion of \SubSubSection{Introduction to Accounting
Architecture}.

Additionally, the course should provide practical experience with the
components of the technology section by encouraging students to build a
computer, install an operating system, create a virtual machine,
administrate and query a database, format a partition, host a website,
use MapReduce on unstructured data, manage a private cloud, implement a
firewall and set up a local-area network. Not all of these experiences
have equal value, but each exposes the students to technological
components of most enterprise-grade information systems.